Новая реализация updater2.exe с защитой от del-bomb.

+++ updater2/Program2.cs
@@ -1,6 +1,8 @@
-﻿using Serilog;
+﻿#define DEBUG
+using Serilog;
 using Serilog.Formatting.Compact;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Reflection;
@@ -19,32 +21,58 @@ namespace updater
 
             InitLogger();
             _log = Log.Logger;
-            _log.Information("Start application 'updater2', version: {ver}", FileVersionInfo.GetVersionInfo(Assembly.GetExecutingAssembly().Location).ProductVersion);
+            var productVersion = FileVersionInfo.GetVersionInfo(Assembly.GetExecutingAssembly().Location).ProductVersion;
+            _log.Information("Start application '{app}', version: {ver}", "updater2", productVersion);
             Thread.Sleep(3000);
 
-            var dir = Directory.GetCurrentDirectory();
-_log.Information("Current dir = '{dir}'", dir);
-            var dirSource = new DirectoryInfo(dir);
-            var curDirName = dirSource.Name;
-_log.Information("Current dir = '{curDirName}'", curDirName);
-            var dirTarget = new DirectoryInfo(dir).Parent;
-_log.Information("Current dirTarget.FullName = '{dirTarget}'", dirTarget.FullName);
+            var exeDir = new DirectoryInfo(Path.TrimEndingDirectorySeparator(AppDomain.CurrentDomain.BaseDirectory));
+            var exeDirFullName = exeDir.FullName;
+            _log.Debug("exeDirFullName = '{exeDirFullName}'", exeDirFullName);
+
+            var curDir = new DirectoryInfo(Directory.GetCurrentDirectory());
+            var curDirFullName = curDir.FullName;
+            _log.Debug("curDirFullName = '{curDirFullName}'", curDirFullName);
+
+            _log.Debug("? '{curDirFullName}' == '{exeDirParentFullName}'", curDir.FullName, exeDir.Parent.FullName);
+            if (curDir.FullName != exeDir.Parent.FullName)
+            {
+                _log.Fatal("Sanity check #1 failed! Current working directory '{curDirFullName}' must be parent for execution directory '{exeDirFullName}'", curDir.FullName, exeDir.FullName);
+                throw new InvalidOperationException($"Sanity check #1 failed! Current working directory '{curDir.FullName}' must be parent for execution directory '{exeDir.FullName}'");
+            }
+            _log.Debug("? '{exeDirName}' == '{productVersion}'", exeDir.Name, productVersion);
+            if (exeDir.Name != productVersion)
+            {
+                _log.Fatal("Sanity check #2 failed! Execution directory name '{exeDirName}' must be equal assembly version {productVersion}", exeDir.Name, productVersion);
+                throw new InvalidOperationException($"Sanity check #2 failed! Execution directory name '{exeDir.Name}' must be equal assembly version {productVersion}");
+            }
+            var dirSource = exeDir;
+            var dirTarget = curDir; // = exeDir.Parent
 
             try
             {
-                // Remove old version in ../ -Exclude Logs,{packages[0].LatestVersion}
-                _log.Information("Remove old files in dir '{dirTarget}'", dirTarget.FullName); // FIXME ! BUG WAS HERE ! Remove all files & dirs in 'Z:\' !
-/*
+                // Remove old version of app in directory 'exeDir.Parent.FullName'
+                _log.Information("Remove old files in dir '{dirTargetFullName}'", dirTarget.FullName);
+                var excludeDirsList = new List<string>() {
+                    productVersion,
+                    @"Logs"
+                };
                 foreach (var delDir in dirTarget.GetDirectories())
-                    if (delDir.Name != "Logs" || delDir.Name != curDirName) delDir.Delete(true);
+                    if (excludeDirsList.Contains(delDir.Name))
+                    {
+                        _log.Debug($"SKIP dir '{delDir.Name}'");
+                    }
+                    else
+                    {
+                        _log.Debug($"Delete dir '{delDir.Name}'");
+                        delDir.Delete(true);
+                    };
                 foreach (var file in dirTarget.GetFiles())
                     file.Delete();
-*/
                 Thread.Sleep(1000);
 
                 // Copy all files and directories from 'sourceDirectory' (./*.*) to 'targetDirectory' (../).
-                _log.Information("Copy files from dir '{dirSource}' into dir '{dirTarget}'", dirSource.FullName, dirTarget.FullName);
-//                CopyFilesRecursively(dirSource, dirTarget);
+                _log.Information("Copy files from dir '{dirSourceFullName}' into dir '{dirTargetFullName}'", dirSource.FullName, dirTarget.FullName);
+                CopyFilesRecursively(dirSource, dirTarget);
                 Thread.Sleep(3000);
             }
             catch (Exception e)
@@ -53,15 +81,26 @@ _log.Information("Current dirTarget.FullName = '{dirTarget}'", dirTarget.FullNam
             }
 
             _log.Information("Restart application!");
-            Process.Start(Path.Combine(dirTarget.FullName, "updater.exe"));
-            Process.GetCurrentProcess().Kill(); // Stop updater2.exe
+            var processInfo = new ProcessStartInfo
+            {
+                WorkingDirectory = dirTarget.FullName,
+                FileName = Path.Combine(dirTarget.FullName, "updater.exe"),
+                Arguments = "",
+                CreateNoWindow = false,
+                UseShellExecute = false,
+                RedirectStandardError = false,
+                RedirectStandardOutput = false,
+                RedirectStandardInput = false
+            };
+            Process.Start(processInfo); // Start new version of 'updater.exe'
+            Process.GetCurrentProcess().Kill(); // Stop 'updater2.exe'
         }
 
         private static void InitLogger()
         {
             var formatter = new CompactJsonFormatter();
 
-            string logPath = $"{Directory.GetCurrentDirectory()}/Logs/";
+            string logPath = Path.Combine(Directory.GetCurrentDirectory(), $"Logs{Path.DirectorySeparatorChar}");
 
             var logger = new LoggerConfiguration()
                 .WriteTo.Console()
